refactor(dispatch): isolate lightweight runtime helpers
