+++ src/tools/dispatch.ts
@@ -81,6 +81,8 @@ import {
 	stringArg,
 	timeoutMsArg,
 } from "./dispatch-arguments.js";
+import type { DispatchBackgroundRegistry } from "./dispatch-background.js";
+import { assistantTextFromEvent } from "./dispatch-event-text.js";
 import {
 	DISPATCH_PLAN_PREPARATION_ERROR_ARGUMENT,
 	describeDispatchPlan,
@@ -108,63 +110,14 @@ import {
 	workerTextNonEvidenceNotices,
 } from "./worker-evidence.js";
 
+export type {
+	DispatchBackgroundControl,
+	DispatchBackgroundOutcome,
+	DispatchBackgroundRegistry,
+} from "./dispatch-background.js";
+
 const TRUNCATION_MARKER = "\n[agent output truncated]";
 const THINKING_LEVELS = ["off", "minimal", "low", "medium", "high", "xhigh", "max"] as const;
-
-/**
- * Result of an operator-initiated attach to detach conversion. The message is
- * the line the keypress feedback renders, so a refusal names the topology that
- * refused rather than reading as a dropped key.
- */
-export type DispatchBackgroundOutcome = { ok: true; message: string } | { ok: false; message: string };
-
-export interface DispatchBackgroundControl {
-	/** Tool call the control belongs to; one attached dispatch owns one control. */
-	toolCallId: string;
-	/** Operator-facing name of the call, used in both outcome messages. */
-	label: string;
-	convert(): DispatchBackgroundOutcome;
-}
-
-/**
- * Live attached dispatches that the operator may send to the background. An
- * attached `dispatch` call registers one control for as long as it awaits its
- * runs; the TUI keybinding fires the newest one. Absent from a deps bundle
- * (workers, headless) nothing registers and attached dispatch behaves exactly
- * as it did before.
- */
-export interface DispatchBackgroundRegistry {
-	/** Returns the deregistration handle; calling it more than once is harmless. */
-	register(control: DispatchBackgroundControl): () => void;
-	/** Fire the newest still-registered control. */
-	backgroundNewest(): DispatchBackgroundOutcome;
-	/** How many attached dispatches are currently registered. */
-	size(): number;
-}
-
-export function createDispatchBackgroundRegistry(): DispatchBackgroundRegistry {
-	// Map insertion order is registration order, and the newest attached dispatch
-	// is the one the operator just watched start, so the keypress takes the last
-	// entry rather than asking a UI projection which segment is newest.
-	const controls = new Map<string, DispatchBackgroundControl>();
-	return {
-		register(control) {
-			controls.set(control.toolCallId, control);
-			return () => {
-				if (controls.get(control.toolCallId) === control) controls.delete(control.toolCallId);
-			};
-		},
-		backgroundNewest() {
-			let newest: DispatchBackgroundControl | undefined;
-			for (const control of controls.values()) newest = control;
-			if (newest === undefined) {
-				return { ok: false, message: "background: no attached dispatch is running" };
-			}
-			return newest.convert();
-		},
-		size: () => controls.size,
-	};
-}
 
 /**
  * One-shot latch an attached executor races against its runs. Firing it never
@@ -470,10 +423,6 @@ export function createDispatchRunEventRegistry(): DispatchRunEventRegistry {
 
 function rawAssistantTextFromEvent(event: unknown): string {
 	return durableAssistantTextFromEvent(event);
-}
-
-export function assistantTextFromEvent(event: unknown): string {
-	return rawAssistantTextFromEvent(event).trim();
 }
 
 function normalizedAssistantText(summary: EventSummary): string {
